Removes the `bevy_feathers` dependency on `bevy_winit`.
This required moving out the `CustomCursor` types from `bevy_winit`. The natural place to put them seemed to be `bevy_window` but `bevy_asset` depends on `bevy_window` so the custom cursor image asset creates a circular dependency. Went to move `CustomCursorImage`  into `bevy_image` instead but that just creates the same cycle with an extra step. Couldn't think of any other sensible alternative so I went to the extremely silly lengths of adding an entire new `bevy_window_cursor` crate.

The `bevy_window` and `bevy_winit` types that are used to set the cursor appearance: `SystemCursorIcon`,  `CursorIcon`, `CustomCursor`, and `CustomCursorImage` have been moved to the new crate.
The plugin `WindowCursorPlugin` just exists to init the `CustomIcon` resource.
