+++ crates/bevy_feathers/Cargo.toml
@@ -15,6 +15,7 @@ bevy_app = { path = "../bevy_app", version = "0.17.0-dev" }
 bevy_asset = { path = "../bevy_asset", version = "0.17.0-dev" }
 bevy_color = { path = "../bevy_color", version = "0.17.0-dev" }
 bevy_core_widgets = { path = "../bevy_core_widgets", version = "0.17.0-dev" }
+bevy_window_cursor = { path = "../bevy_window_cursor", version = "0.17.0-dev" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.17.0-dev" }
 bevy_input_focus = { path = "../bevy_input_focus", version = "0.17.0-dev" }
 bevy_log = { path = "../bevy_log", version = "0.17.0-dev" }
@@ -35,7 +36,6 @@ accesskit = "0.21"
 
 [features]
 default = []
-# custom_cursor = ["bevy_winit/custom_cursor"]
 
 [lints]
 workspace = true
