+++ crates/bevy_feathers/src/controls/checkbox.rs
@@ -74,7 +74,7 @@ pub fn checkbox<C: SpawnableList<ChildOf> + Send + Sync + 'static, B: Bundle>(
         },
         CheckboxFrame,
         Hovered::default(),
-        EntityCursor::System(bevy_window::SystemCursorIcon::Pointer),
+        EntityCursor::System(bevy_window_cursor::SystemCursorIcon::Pointer),
         TabIndex(0),
         ThemeFontColor(tokens::CHECKBOX_TEXT),
         InheritableFont {
