+++ crates/bevy_image/Cargo.toml
@@ -11,6 +11,8 @@ keywords = ["bevy"]
 [features]
 default = ["bevy_reflect"]
 
+custom_cursor = []
+
 # bevy_reflect can't optional as it's needed for TypePath
 # this feature only control reflection in bevy_image
 bevy_reflect = ["bevy_math/bevy_reflect"]
