+++ crates/bevy_image/src/lib.rs
@@ -25,6 +25,8 @@ pub use self::serialized_image::*;
 mod basis;
 #[cfg(feature = "compressed_image_saver")]
 mod compressed_image_saver;
+#[cfg(feature = "custom_cursor")]
+mod custom_cursor;
 #[cfg(feature = "dds")]
 mod dds;
 mod dynamic_texture_atlas_builder;
@@ -40,6 +42,8 @@ mod texture_atlas_builder;
 
 #[cfg(feature = "compressed_image_saver")]
 pub use compressed_image_saver::*;
+#[cfg(feature = "custom_cursor")]
+pub use custom_cursor_image::*;
 #[cfg(feature = "dds")]
 pub use dds::*;
 pub use dynamic_texture_atlas_builder::*;
