+++ crates/bevy_internal/src/lib.rs
@@ -33,6 +33,8 @@ pub use bevy_color as color;
 pub use bevy_core_pipeline as core_pipeline;
 #[cfg(feature = "bevy_core_widgets")]
 pub use bevy_core_widgets as core_widgets;
+#[cfg(feature = "bevy_window_cursor")]
+pub use bevy_window_cursor as cursor;
 #[cfg(feature = "bevy_dev_tools")]
 pub use bevy_dev_tools as dev_tools;
 pub use bevy_diagnostic as diagnostic;
