+++ crates/bevy_window_cursor/README.md
@@ -1,0 +1,7 @@
+# Bevy Color
+
+[![License](https://img.shields.io/badge/license-MIT%2FApache-blue.svg)](https://github.com/bevyengine/bevy#license)
+[![Crates.io](https://img.shields.io/crates/v/bevy_window_cursor.svg)](https://crates.io/crates/bevy_window_cursor)
+[![Downloads](https://img.shields.io/crates/d/bevy_window_cursor.svg)](https://crates.io/crates/bevy_window_cursor)
+[![Docs](https://docs.rs/bevy_window_cursor/badge.svg)](https://docs.rs/bevy_window_cursor/latest/bevy_window_cursor/)
+[![Discord](https://img.shields.io/discord/691052431525675048.svg?label=&logo=discord&logoColor=ffffff&cursor=7389D8&labelColor=6A7EC2)](https://discord.gg/bevy)
