+++ release-content/migration-guides/bevy_cursor_crate.md
@@ -1,0 +1,4 @@
+---
+title: bevy_window_cursor crate
+pull_requests: []
+---
